Data.Array.Accelerate.Examples.Internal

Description

This module exposes some internal infrastructure that is useful and common to the accelerate-examples test and benchmark program suite. This is not intended for general consumption.

Options processing

parseArgs

Arguments

:: [OptDescr (config -> config)]
the user option descriptions

-> config
user default option set

-> [String]
header text

-> [String]
footer text

-> IO (config, Options, [String])

Process the command line arguments and return a tuple consisting of the user options structure, accelerate-examples options (including options for criterion and test-framework), and a list of unrecognised command line arguments.

Since criterion and test-framework both bail if they encounter unrecognised options, we run getOpt' ourselves. This means that the error messages might be slightly different.

Any command line arguments following a "--" are not processed, but are included as part of the unprocessed arguments returned on output.

data Options

optBackend :: forall cat. ArrowApply cat => Lens cat Options Backend

optTest :: forall cat. ArrowApply cat => Lens cat Options Bool

optBenchmark :: forall cat. ArrowApply cat => Lens cat Options Bool

optCodespeed :: forall cat. ArrowApply cat => Lens cat Options (Maybe String)

optHostname :: forall cat. ArrowApply cat => Lens cat Options String

optVariant :: forall cat. ArrowApply cat => Lens cat Options String

optHelp :: forall cat. ArrowApply cat => Lens cat Options Bool

optCriterion :: forall cat. ArrowApply cat => Lens cat Options Config

optTestFramework :: forall cat. ArrowApply cat => Lens cat Options Config

module System.Console.GetOpt

Executing programs

run :: Arrays a => Backend -> Acc a -> a

Execute Accelerate expressions

run1 :: (Arrays a, Arrays b) => Backend -> (Acc a -> Acc b) -> a -> b

run2 :: (Arrays a, Arrays b, Arrays c) => Backend -> (Acc a -> Acc b -> Acc c) -> a -> b -> c

data Backend

The set of backends available to execute the program.

Constructors

Interpreter

CUDA

CPU

PTX

defaultBackend :: Backend

availableBackends :: (options :-> Backend) -> [OptDescr (options -> options)]

concurrentBackends :: Backend -> Maybe Int

Criterion benchmarking

runBenchmarks :: Options -> [String] -> [Benchmark] -> IO ()

Run the given benchmarks, if we have enabled benchmark mode.

nf :: NFData b => (a -> b) -> a -> Benchmarkable

whnf :: (a -> b) -> a -> Benchmarkable

nfIO :: NFData a => IO a -> Benchmarkable

whnfIO :: IO a -> Benchmarkable

data Benchmark :: *

env :: NFData env => IO env -> (env -> Benchmark) -> Benchmark

bench :: String -> Benchmarkable -> Benchmark

bgroup :: String -> [Benchmark] -> Benchmark

Interactive/GUI

runInteractive :: Options -> [String] -> IO () -> IO ()

If accelerate-examples is configured to enable GUI programs, and we are not in benchmarking mode, then execute the given action.

Test-framework runner

data Test :: *

runTests :: Options -> [String] -> [Test] -> IO ()

Run the given tests, if enabled.

QuickCheck

data Property :: *

testProperty :: Testable a => TestName -> a -> Test

testGroup :: TestName -> [Test] -> Test

(~=?) :: (Similar a, Show a) => a -> a -> Property infix 1

Short hand for a test case that asserts similarity, with the actual value on the right hand side and the expected value on the left.

(~?=) :: (Similar a, Show a) => a -> a -> Property infix 1

Short hand for a test case that asserts similarity, with the actual value on the left hand side and the expected value on the right.

HUnit

type Assertion = IO ()

testCase :: TestName -> Assertion -> Test

assertEqual

Arguments

:: (Similar a, Show a)

=> a
The expected value

-> a
The actual value

-> Assertion

Assert that the specified actual value is equal-ish to the expected value. If we are in verbose mode, the output message will contain the expected and actual values.

Similarity

class Similar a where

Minimal complete definition

Nothing

Methods

(~=) :: a -> a -> Bool infix 4

Resource monitoring

beginMonitoring :: IO ()

Launch a monitoring server that will collect statistics on the running application. This should be called as soon as the application starts. The program will need to be run with the RTS option -T.

Generating random arrays

type (:~>) sh e = sh -> GenIO -> IO e

A PNRG from indices to variates

uniform :: (Shape sh, Elt e, Variate e) => sh :~> e

Uniformly distributed random variates.

uniformR :: (Shape sh, Elt e, Variate e) => (e, e) -> sh :~> e

Uniformly distributed random variates in a given range.

randomArray :: (Shape sh, Elt e) => (sh :~> e) -> sh -> Array sh e

Generate an array of random values using the supplied generator function. The generator for variates is initialised with a fixed seed.

randomArrayWithSeed :: (Shape sh, Elt e) => Seed -> (sh :~> e) -> sh -> Array sh e

Generate an array of random values using a supplied generator function and seed value.

randomArrayWithSystemRandom :: forall sh e. (Shape sh, Elt e) => (sh :~> e) -> sh -> IO (Array sh e)

Generate an array of random values using a supplied generator function, initialised with the system's source of pseudo-random numbers.

TODO: find a way to do this directly, without going via save/restore.

Miscellaneous

showFFloatSIBase :: RealFloat a => Maybe Int -> a -> a -> ShowS

Show a floating point number in scientific notation with a specific base.